+++ packages/ethereum-indexer-utils/src/indexer.ts
@@ -1,6 +1,16 @@
-import {Abi, LastSync} from 'ethereum-indexer';
+import {Abi, LastSync, ProcessorContext, simple_hash} from 'ethereum-indexer';
 import {filterOutFieldsFromObject} from './javascript';
 
 export function formatLastSync<ABI extends Abi>(lastSync: LastSync<ABI>): any {
 	return filterOutFieldsFromObject(lastSync, ['_rev', '_id', 'batch']);
+}
+
+
+export function contextFilenames(context: ProcessorContext<Abi, any>) {
+	const configHash = 'config' in context && context.config ? simple_hash(context.config) : undefined;
+	const networkString = `${context.source.chainId }${(context.source.chainId == '1337' || context.source.chainId == '31337') && context.source.genesisHash ? `-${context.source.genesisHash}`: ''}`
+	const prefix = `${networkString}${configHash ? `-${configHash}`: ``}${context.version ? `-${context.version}`: ``}`;
+	const stateFile = `${prefix}-state.json`;
+	const lastSyncFile = `${prefix}-lastSync.json`;
+	return {stateFile,lastSyncFile}
 }